+++ WordPress/Tests/VIP/TimezoneChangeUnitTest.inc
@@ -1,6 +1,6 @@
 <?php
 
+// @codingStandardsChangeSetting WordPress.VIP.TimezoneChange exclude something
 date_default_timezone_set( 'Foo/Bar' ); // Bad.
 
-$date = new DateTime();
-$date->setTimezone( new DateTimeZone( 'America/Toronto' ) ); // Yay!
+// @codingStandardsChangeSetting WordPress.VIP.TimezoneChange exclude false
