+++ WordPress/Tests/WP/TimezoneChangeUnitTest.inc
@@ -1,0 +1,6 @@
+<?php
+
+date_default_timezone_set( 'Foo/Bar' ); // Bad.
+
+$date = new DateTime();
+$date->setTimezone( new DateTimeZone( 'America/Toronto' ) ); // Yay!
