+++ WordPress/ruleset.xml
@@ -152,4 +152,14 @@
 		<severity>0</severity>
 	</rule>
 
+	<!-- Prevent deprecation notice when the sniff is not explicitely included. -->
+	<rule ref="WordPress.VIP.TimezoneChange.DeprecatedSniff">
+		<severity>0</severity>
+	</rule>
+
+	<!-- Prevent duplicate messages from deprecated sniff. -->
+	<rule ref="WordPress.VIP.TimezoneChange.timezone_change_date_default_timezone_set">
+		<severity>0</severity>
+	</rule>
+
 </ruleset>
